v1.3
Created the GenerateMenu function, which reads in a dictionary of menu
items, and turns them into DirectoryObjects and returns them in a
ObjectContainer. Used this function everywhere I could to cut down on
repetition and lines of code. Merged the code from the MLBrowseVideos.py
file into MLCommon.py, cutting down on a file. Added in sort orders for
members. Moved the tag search to the main menu. General code cleanup,
trying to make everything uniform.